Fixed issues pointed out by static analysis.

+++ src/libxsmm_cpuid_x86.c
@@ -223,8 +223,8 @@ LIBXSMM_API int libxsmm_cpuid_x86(libxsmm_cpuid_info* info)
             ? "highly " : NULL));
         if (NULL != compiler_support) {
           const char *const name = libxsmm_cpuid_name( /* exclude MIC when running on Core processors */
-            (((LIBXSMM_X86_AVX512_MIC == LIBXSMM_MAX_STATIC_TARGET_ARCH) ||
-              (LIBXSMM_X86_AVX512_KNM == LIBXSMM_MAX_STATIC_TARGET_ARCH)) && (LIBXSMM_X86_AVX512_CORE <= feature_cpu))
+            (((int)(LIBXSMM_X86_AVX512_MIC == LIBXSMM_MAX_STATIC_TARGET_ARCH) ||
+              (int)(LIBXSMM_X86_AVX512_KNM == LIBXSMM_MAX_STATIC_TARGET_ARCH)) && (LIBXSMM_X86_AVX512_CORE <= feature_cpu))
               ? LIBXSMM_X86_AVX2 : LIBXSMM_MAX_STATIC_TARGET_ARCH);
           fprintf(stderr, "LIBXSMM WARNING: %soptimized non-JIT code paths are limited to \"%s\"!\n", compiler_support, name);
         }
